Add user online status UI and widget refactor Wire `userStatus` events from WebSocket service/controller into `MainWindow`, track per-user online state, and update the interlocutor header to show online/offline text with status-aware styling. Extract inline preview/avatar overlay widget classes from `mainwindow.cpp` into dedicated `widgets/` files and update call sites to use the new APIs. Also skip attachment re-downloads when local files already exist and apply small consistency cleanups (`quint64` userId type, connect/style formatting).

+++ src/services/userinfoservice.h
@@ -20,7 +20,7 @@ struct ParsedFoundUsersObject
     QString            nickname;                          //!< Никнейм (отображаемое имя)
     QString            lastSeen;                          //!< Временная метка последней активности
     QString            avatarFileUrl;                     //!< Ссылка на аватар
-    unsigned long long userId = ULONG_LONG_MAX;           //!< Идентификатор пользователя
+    quint64            userId = ULONG_LONG_MAX;           //!< Идентификатор пользователя
 };
 
 /**
